+++ src/senaite/ast/browser/results.py
@@ -24,6 +24,7 @@ from collections import OrderedDict
 from bika.lims import api
 from bika.lims.browser.analyses import AnalysesView
 from bika.lims.interfaces import IVerified
+from bika.lims.utils import get_image
 from bika.lims.utils import get_link
 from plone.memoize import view
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
@@ -31,6 +32,9 @@ from senaite.ast import is_installed
 from senaite.ast import messageFactory as _
 from senaite.ast import utils
 from senaite.ast.config import AST_POINT_OF_CAPTURE
+from senaite.ast.config import IDENTIFICATION_KEY
+from senaite.ast.utils import get_ast_analyses
+from senaite.ast.i18n import translate as t
 from senaite.core.browser.viewlets.sampleanalyses import LabAnalysesViewlet
 
 
@@ -42,9 +46,24 @@ class ASTAnalysesViewlet(LabAnalysesViewlet):
     capture = AST_POINT_OF_CAPTURE
 
     def available(self):
-        """Returns true if senaite.ast is installed
+        """Returns true if senaite.ast is installed and the sample contains
+        at least one sensitivity testing analysis or the microorganism
+        identification analysis is present
         """
-        return is_installed()
+        if not is_installed():
+            return False
+
+        # does this sample has the identification analysis?
+        analyses = self.context.getAnalyses(getKeyword=IDENTIFICATION_KEY)
+        if analyses:
+            return True
+
+        # does this have sensitivity testing analyses?
+        ast_analyses = get_ast_analyses(self.context)
+        if ast_analyses:
+            return True
+
+        return False
 
 
 class ManageResultsView(AnalysesView):
@@ -166,6 +185,9 @@ class ManageResultsView(AnalysesView):
 
             # This interim will be displayed as readonly mode, display text
             text = utils.get_interim_text(interim_field, default="")
+            if interim_field.get("status_rejected", False):
+                icon = get_image('warning.png', title=t(_("Not tested")))
+                text = "{}{}".format(text, icon)
             item["replace"][keyword] = text or "&nbsp;"
 
     def folderitems(self):